requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
diff-match-patch==20121119
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.4
django-bootstrap3==9.0.0
django-import-export==0.5.1
et-xmlfile==1.0.1
gunicorn==19.7.1
idna==2.6
jdcal==1.3
odfpy==1.3.5
openpyxl==2.4.8
psycopg2==2.7.3.1
pytz==2017.2
PyYAML==3.12
requests==2.18.4
static3==0.7.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.22
xlrd==1.1.0
xlwt==1.3.0